Final version-exculding audio

+++ Assets/Scipts/damage.cs
@@ -9,7 +9,7 @@ public class damage : MonoBehaviour
     [SerializeField] damageType type; // Type of damage behavior
     [SerializeField] Rigidbody rb;
 
-    [SerializeField] int damageAmount;
+    [SerializeField] float damageAmount;
     [SerializeField] float damageRate;
     [SerializeField] int speed;
     [SerializeField] int destroyTime;
